Rename auxTel to LATISS
The telescope is AuxTel but the camera on that telscope is
called LATISS.
