+++ config/latiss/processCcd.py
