+++ config/latiss/processStar.py
