+++ data/input/latiss/CALIB/bias/2018-09-21/bias-det000_2018-09-21.fits
