+++ data/input/latiss/raw/2018-09-20/2018092000065-det000.fits
