+++ policy/.gitignore
@@ -1,4 +1,4 @@
-auxTel.yaml
+latiss.yaml
 imsim.yaml
 lsstCam.yaml
 phosim.yaml
