+++ policy/SConscript
@@ -14,10 +14,11 @@ for name in ("PYTHONPATH", "LD_LIBRARY_PATH",
 # we may need an explicit library load path specified in the command
 libpathstr = lsst.sconsUtils.utils.libraryLoaderEnvironment()
 
-# We always run these commands with an explicit python rather than relying on the shebang
+# We always run these commands with an explicit python rather than relying on
+# the shebang
 python = "{} python".format(libpathstr)
 
-for camera in ["auxTel", "lsstCam", "imsim", "phosim", "ts8", "ucd", "ts3", "comCam"]:
+for camera in ["latiss", "lsstCam", "imsim", "phosim", "ts8", "ucd", "ts3", "comCam"]:
     # Invoke the bin.src variant so that we do not depend on the shebang
     # target having been run
     commandInst = env.Command(f"{camera}.yaml", "",
