+++ python/lsst/obs/lsst/auxTel.py
@@ -20,107 +20,11 @@
 # the GNU General Public License along with this program.  If not,
 # see <http://www.lsstcorp.org/LegalNotices/>.
 #
-import os.path
-import re
-import lsst.log
-from . import LsstCamMapper, LsstCamMakeRawVisitInfo
-from .ingest import LsstCamParseTask
-from .translators import LsstAuxTelTranslator
 
-__all__ = ["AuxTelMapper", "AuxTelParseTask"]
+"""Backwards compatibility shim to allow gen2 repos to continue to
+work with the legacy AuxTelMapper name.
+"""
 
+from .latiss import LatissMapper as AuxTelMapper
 
-class AuxTelMakeRawVisitInfo(LsstCamMakeRawVisitInfo):
-    """Make a VisitInfo from the FITS header of a raw image."""
-    metadataTranslator = LsstAuxTelTranslator
-
-
-class AuxTelMapper(LsstCamMapper):
-    """The Mapper for the auxTel camera."""
-
-    MakeRawVisitInfoClass = AuxTelMakeRawVisitInfo
-
-    _cameraName = "auxTel"
-    yamlFileList = ["auxTel/auxTelMapper.yaml"] + list(LsstCamMapper.yamlFileList)
-
-    def _extractDetectorName(self, dataId):
-        return f"{LsstAuxTelTranslator.DETECTOR_GROUP_NAME}_{LsstAuxTelTranslator.DETECTOR_NAME}"
-
-    def _computeCcdExposureId(self, dataId):
-        """Compute the 64-bit (long) identifier for a CCD exposure.
-
-        Parameters
-        ----------
-        dataId : `dict`
-            Data identifier including dayObs and seqNum.
-
-        Returns
-        -------
-        id : `int`
-            Integer identifier for a CCD exposure.
-        """
-        if len(dataId) == 0:
-            return 0                    # give up.  Useful if reading files without a butler
-
-        if 'visit' in dataId:
-            visit = dataId['visit']
-        else:
-            visit = LsstAuxTelTranslator.compute_exposure_id(dataId['dayObs'], dataId["seqNum"])
-
-        if "detector" in dataId:
-            detector = dataId["detector"]
-            if detector != 0:
-                lsst.log.Log.getLogger("AuxTelMapper").warn("Got detector %d for AuxTel when it should"
-                                                            " always be 0", detector)
-        else:
-            detector = 0
-
-        return LsstAuxTelTranslator.compute_detector_exposure_id(visit, detector)
-
-
-class AuxTelParseTask(LsstCamParseTask):
-    """Parser suitable for auxTel data.
-    """
-
-    _mapperClass = AuxTelMapper
-    _translatorClass = LsstAuxTelTranslator
-
-    def translate_seqNum(self, md):
-        """Return the sequence number.
-
-        Parameters
-        ----------
-        md : `~lsst.daf.base.PropertyList` or `~lsst.daf.base.PropertySet`
-            Image metadata.
-
-        Returns
-        -------
-        seqnum : `int`
-            The sequence number identifier valid within a day.
-        """
-
-        if "SEQNUM" in md:
-            return md.getScalar("SEQNUM")
-        #
-        # Oh dear.  Extract it from the filename
-        #
-        seqNum = 0
-        for k in ("IMGNAME", "FILENAME"):
-            if k not in md:
-                continue
-            name = md.getScalar(k)           # e.g. AT-O-20180816-00008
-            # Trim trailing extensions
-            name = os.path.splitext(name)[0]
-
-            # Want final digits
-            mat = re.search(r"(\d+)$", name)
-            if mat:
-                seqNum = int(mat.group(1))    # 00008
-                break
-
-        if seqNum == 0:
-            logger = lsst.log.Log.getLogger('obs.lsst.AuxTelParseTask')
-            logger.warn(
-                'Could not determine sequence number. Assuming %d ', seqNum)
-
-        return seqNum
+__all__ = ["AuxTelMapper"]
