+++ python/lsst/obs/lsst/translators/__init__.py
@@ -13,6 +13,6 @@ from .phosim import *
 from .ts3 import *
 from .ts8 import *
 from .imsim import *
-from .auxTel import *
+from .latiss import *
 from .lsst_ucdcam import *
 from .comCam import *
